requirements.txt
netmiko
flask==3.0.3
pillow
pyserial
Werkzeug==3.0.4
Flask-SQLAlchemy==2.5.1
celery 
redis
flask-socketio